+++ index.html
@@ -164,6 +164,7 @@
       </div>
     </section>
   </template>
+  <script src="main.js" type="module"></script>
 
 </body>
 </html>
